feat: refactoring template (#105)
* dynamic source map

* change gulp file path

* cp test dir
