--- templates/default/lib/index.spec.ts
@@ -1,1 +1,0 @@
-test.todo('Add test');
